Remove un-needed 'min' value in editor container (#586)
The height and width boxes were being bound to a lower value based on the ratio which was preventing users from inputting certain values that were lower than the ratio calculated from the other field. This change sets the min to be 1, and lets the other value be calculated based on the ratio as normal. Fixes #576